\name{spsextante.helloworld}
\alias{spsextante.helloworld}
\title{Function to say hello from a Java class}
\description{
  This function executes a method of a Java class
  defined in a JAR inside the R package.
}
\usage{
  helloJavaWorld()
}
\details{
  Details are given in the package vignette.
}
\value{
  The string "Hello Java World!" is returned.
}
\author{Dominik Fuchs}
\seealso{\link[spsextante:../javadoc/index]{The Java API of the package}}
\examples{
  spsextante.helloworld()
}
\keyword{manip}
